Adds unit tests for Happiness decrease and Hunger increase over time

+++ MediatonicApi/Models/Animal.cs
@@ -5,8 +5,8 @@ namespace MediatonicApi.Models
     {
         public uint Id { get; set; }
         public string TypeName { get; set; }
-        public uint HungerPerSecond { get; set; }
-        public uint SadnessPerSecond { get; set; }
+        public double HungerPerSecond { get; set; }
+        public double SadnessPerSecond { get; set; }
 
         public uint MaxHunger { get; set; }
         public int MinHappiness { get; set; }
